Bug 513337 - NPE in ProcessStreamConnectionProvider.hashCode
This happens for example when "node" is not installed, then
CSSLanguageServer adds a null string to the list of commands.

This patch adds a way for providers to be flagged as "invalid" and not
used at all so that they are not even considered to be added to the
"connectionsInfo", which means hashCode will never be called. This
approach will prevent a lot of other NPEs such as calling
ProcessStreamConnectionProvider.stop but the process is null.

But even then, any exception in ProcessStreamConnectionProvider.hashCode
or an overridden version of it will prevent other servers to load, so I
think the part of the code in the registry calling it should be
safe-guarded with a SafeRunner.

Change-Id: Idb16e7a91e7ca1b35fce03c8be87fa3c87fb08c2
Signed-off-by: Marc-Andre Laperle <marc-andre.laperle@ericsson.com>